+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/DispatcherServletInitializer.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.asis;
 
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
