+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.asis;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
