+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/HandlerAdapters.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.asis;
 
 import java.util.List;
